Parse <continue/> without making assumptions on its inner attributes
The "modern" continuation mechanism (as opposed to `rawcontinue`) does
not require API consumers to parse, select and process such attributes.
Instead, passing them all on as they are to the next request should be
enough. See also:

https://www.mediawiki.org/wiki/API:Query#Continuing_queries

The bug solved here is linked to the erroneous assumption that parsing
`<prefix>continue` for list-type requests using the supplied prefix
would be enough. This does not necessarily hold true for generator-based
content queries, for instance:

`prop=revisions&rvprop=content&generator=embeddedin&geititle=Template:Foo`

shall return the following if the number of results *and* `geilimit`
exceed `slowmax` (500 for bots):

`<continue rvcontinue="..." continue="..." />`

This behavior is due to the fact that results of content queries
originating from a generator list are supplied in batches of `slowmax`.
While consumers that benefit from higher limits usually retrieve up to
5000 results on a single `list=...` request, when it comes to fetching
the wikitext with a generator, the MW API will split it in 10 batches of
500 items with a continuation element as shown above. Starting from the
*second* sequence of batches (from the 5001 result onwards), we'll see a
`geicontinue` attribute, as expected by the previous implementation,
*along with* the always present `rvcontinue`.